makes logout its own container and updates snapshots
